+++ backend/src/category/category.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateCategoryDto } from './dto/create-category.dto';
 import { UpdateCategoryDto } from './dto/update-category.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -9,6 +13,9 @@ export class CategoryService {
   constructor(private readonly prisma: PrismaService) {}
 
   async create(createCategoryDto: CreateCategoryDto) {
+    if (!createCategoryDto.name)
+      throw new BadRequestException('Missing name of category');
+
     const newCategory = await this.prisma.category.create({
       data: {
         name: createCategoryDto.name,
@@ -26,10 +33,13 @@ export class CategoryService {
     const category = await this.prisma.category.findUnique({
       where: { id },
     });
+    if (!category) throw new NotFoundException("The category doesn't exist");
     return Category.fromPrisma(category);
   }
 
   async update(id: string, updateCategoryDto: UpdateCategoryDto) {
+    const category = await this.getById(id);
+    if (!category) throw new NotFoundException("The category doesn't exist");
     return this.prisma.category.update({
       where: { id },
       data: updateCategoryDto,
@@ -37,6 +47,8 @@ export class CategoryService {
   }
 
   async remove(id: string) {
+    const category = await this.getById(id);
+    if (!category) throw new NotFoundException("The category doesn't exist");
     return this.prisma.category.delete({ where: { id } });
   }
 }
