+++ backend/src/user/user.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { User } from './entities/user.entity';
@@ -9,6 +13,15 @@ export class UserService {
   constructor(private readonly prisma: PrismaService) {}
 
   async create(createUserDto: CreateUserDto) {
+    if (
+      !createUserDto.email ||
+      !createUserDto.firstName ||
+      !createUserDto.lastName ||
+      !createUserDto.password ||
+      !createUserDto.role
+    )
+      throw new BadRequestException('The user object is invalid');
+
     const newUser = await this.prisma.user.create({
       data: createUserDto,
     });
@@ -25,6 +38,7 @@ export class UserService {
     const user = await this.prisma.user.findUnique({
       where: { id },
     });
+    if (!user) throw new NotFoundException("The user doesn't exist");
     return User.fromPrisma(user);
   }
 
@@ -32,14 +46,19 @@ export class UserService {
     const user = await this.prisma.user.findUnique({
       where: { email },
     });
+    if (!user) throw new NotFoundException("The user doesn't exist");
     return User.fromPrisma(user);
   }
 
   async update(id: string, updateUserDto: UpdateUserDto) {
+    const user = this.getById(id);
+    if (!user) throw new NotFoundException("The user doesn't exist");
     return this.prisma.user.update({ where: { id }, data: updateUserDto });
   }
 
   async remove(id: string) {
+    const user = this.getById(id);
+    if (!user) throw new NotFoundException("The user doesn't exist");
     return this.prisma.user.delete({ where: { id } });
   }
 }
